Add Authors and Books views and services
This commit adds view templates for 'Authors' and 'Books' and corresponding controller classes. It also introduces service layer with services and implementations for Book and Author. These changes setup the structure needed to display lists of 'Authors' and 'Books'. It also includes Thymeleaf dependency in the pom.xml for HTML templates.